[FEATURE] Add initial Vagrant configuration
This adds a Vagrant file and some related Chef artefacts
(such as roles and data bag items) and a Cheffile which
contains links to the required cookbooks.

This is still work in progress, as it lacks documentation
and some refinements - but it generally works if you
know how to use librarian-chef and Vagrant.

More to come.

Change-Id: I92f898686c1f3166a05d1778a3c4a789853ea212